Deployment/requirements.txt
matplotlib==3.5.1
numpy==1.19.5
pandas==1.3.5
Pillow==9.0.1
streamlit==1.7.0
tensorflow==2.8.0
